# Adding Utf8Json to Json Reader performance tests